+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### Version 1.27 [27-09-2024]
+Updated BigWigs profiles (Full season 1 colors)  
+Updated ElvUI blacklist  
+Updated ElvUI profile (Status text dead + offline)  
+Updated ElvUI color low health breakpoint (now 45% was 40%)  
+___
 ### Version 1.26 [20-09-2024]
 Added CVars step to the installer and config panel  
 Added important WeakAura to handle friendly nameplates